+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   * The reputation detail frame had an out of place background
   * [classic] Error when mousing over the skill detail status bar
   * [classic] Spellbook tabs would not react to mouse input
+  * [tbc] The character title dropdown had a misaligned backdrop
 
 
 ## [9.1.5.0] ##